Reorganized frequency serializers
