+++ src/frequencies/mod.rs
@@ -3,6 +3,7 @@ mod frequencies_tests;
 mod frequencies_validation_tests;
 mod models;
 mod validations;
+mod serializer;
 
 pub use errors::InvalidFrequency;
 pub use models::{Frequency, MonthlyDate, NthWeekday, Time};
