+++ src/lib.rs
@@ -1,6 +1,5 @@
 mod frequencies;
 mod recurrences;
-mod serializer;
 mod utils;
 
 pub use frequencies::{Frequency, MonthlyDate, NthWeekday, Time};
